+++ StartButton.cpp
@@ -18,8 +18,8 @@ void StartButton::update()
 
 void StartButton::render(HDC _hdc)
 {
-	int x = lt.x + (size - 100.0f);
-	int y = lt.y + (size - 100.0f);
+	int x = (int)(lt.x + (size - 100.0f));
+	int y = (int)(lt.y + (size - 100.0f));
 
 	COLORREF currnetColor = RGB(0, 0, 0);
 	switch (currentState)
@@ -38,7 +38,7 @@ void StartButton::render(HDC _hdc)
 		size = 100.0f;
 		break;
 	}
-	ImageHandler::textResize(size, _hdc);
+	ImageHandler::textResize((int)size, _hdc);
 
 	const TCHAR* text = _T("Start");
 	
@@ -47,7 +47,7 @@ void StartButton::render(HDC _hdc)
 	
 	lt = { (cam->getWindowSize().x / 2) - textSize.cx / 2, 400.0f };
 	rb = { (cam->getWindowSize().x / 2) + textSize.cx / 2, 400.0f + textSize.cy - 20 };
-	ImageHandler::DrawOutlinedText(_hdc, text, lt.x, lt.y, currnetColor);
+	ImageHandler::DrawOutlinedText(_hdc, text, (int)lt.x, (int)lt.y, currnetColor);
 }
 
 void StartButton::triggered()
